ENH Multiple improvements, cleanups and fixes from PR review
