+++ ci/run_cuml_singlegpu_pytests.sh
@@ -5,3 +5,7 @@
 cd "$(dirname "$(realpath "${BASH_SOURCE[0]}")")"/../python/cuml/cuml/tests
 
 python -m pytest --cache-clear --ignore=dask -m "not memleak" "$@" .
+
+cd "$(dirname "$(realpath "${BASH_SOURCE[0]}")")"/../python/cuml/cuml/tests/experimental/accel
+
+python -m pytest -p cuml.experimental.accel --cache-clear "$@" .
