--- python/cuml/cuml/experimental/accel/annotation.py
@@ -1,47 +1,0 @@
-#
-# Copyright (c) 2024, NVIDIA CORPORATION.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#
-
-from __future__ import annotations
-
-try:
-    import nvtx
-except ImportError:
-
-    class nvtx:  # type: ignore
-        """Noop-stub with the same API as nvtx."""
-
-        push_range = lambda *args, **kwargs: None  # noqa: E731
-        pop_range = lambda *args, **kwargs: None  # noqa: E731
-
-        class annotate:
-            """No-op annotation/context-manager"""
-
-            def __init__(
-                self,
-                message: str | None = None,
-                color: str | None = None,
-                domain: str | None = None,
-                category: str | int | None = None,
-            ):
-                pass
-
-            def __enter__(self):
-                return self
-
-            def __exit__(self, *exc):
-                return False
-
-            __call__ = lambda self, fn: fn  # noqa: E731
